[demo][hello] refactor directory/package name(s)
