+++ demo/hello/.gitignore
@@ -1,0 +1,4 @@
+
+#
+lib/coverage.out
+
