+++ demo/hello/lib/hello_test.go
@@ -1,14 +1,14 @@
-package hello
+package lib
 
 import (
 	"fmt"
 	"testing"
 )
 
-// run tests with; ... % cd hello;  % go test . -v
+// run tests with; ... % cd lib;  % go test . -v
 
 func Test_hello(t *testing.T) {
-	expect := "hello world"
+	expect := "lib world"
 	if Hello(nil) != expect {
 		t.Errorf("failed, expect: %s", expect)
 	}
@@ -19,13 +19,13 @@ func Test_hello(t *testing.T) {
 	}
 
 	kitty := "kitty"
-	expect = fmt.Sprintf("hello %s", kitty)
+	expect = fmt.Sprintf("lib %s", kitty)
 	if Hello(&kitty) != expect {
 		t.Errorf("failed, expect: %s", expect)
 	}
 
 	eleven := 11
-	expect = fmt.Sprintf("hello number %d", eleven)
+	expect = fmt.Sprintf("lib number %d", eleven)
 	if HelloInt(eleven) != expect {
 		t.Errorf("failed, expect: %s", expect)
 	}
@@ -39,8 +39,8 @@ func Test_hello_with_table(t *testing.T) {
 		expectedResults []string
 		msg             string
 	}{
-		{name: "first test", values: []string{"Bob", ""}, expectedResults: []string{"hello Bob", "hello world"}, msg: "first hello test(s)"},
-		{name: "next test", values: []string{"there"}, expectedResults: []string{"hello there"}, msg: "next hello test"},
+		{name: "first test", values: []string{"Bob", ""}, expectedResults: []string{"lib Bob", "lib world"}, msg: "first lib test(s)"},
+		{name: "next test", values: []string{"there"}, expectedResults: []string{"lib there"}, msg: "next lib test"},
 	}
 
 	for _, e := range helloEntries {
